Move models/ submodule to v1/
