+++ micro_sam/multi_dimensional_segmentation.py
@@ -1,7 +1,6 @@
 """Deprecated alias for :mod:`micro_sam.v1.multi_dimensional_segmentation`."""
 import warnings
-
-from micro_sam.v1 import multi_dimensional_segmentation as _v1_module
+import importlib
 
 warnings.warn(
     "'micro_sam.multi_dimensional_segmentation' has moved to "
@@ -12,4 +11,4 @@ warnings.warn(
 
 
 def __getattr__(name):
-    return getattr(_v1_module, name)
+    return getattr(importlib.import_module("micro_sam.v1.multi_dimensional_segmentation"), name)
