+++ test/test_models/test_peft_sam.py
@@ -18,49 +18,49 @@ class TestPEFTSam(unittest.TestCase):
             self.assertEqual(masks.shape, expected_shape)
 
     def test_lora_sam(self):
-        from micro_sam.models.peft_sam import PEFT_Sam, LoRASurgery
+        from micro_sam.v1.models.peft_sam import PEFT_Sam, LoRASurgery
 
         _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=LoRASurgery)
         self._check_output(peft_sam)
 
     def test_fact_sam(self):
-        from micro_sam.models.peft_sam import PEFT_Sam, FacTSurgery
+        from micro_sam.v1.models.peft_sam import PEFT_Sam, FacTSurgery
 
         _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=FacTSurgery)
         self._check_output(peft_sam)
 
     def test_attention_layer_peft_sam(self):
-        from micro_sam.models.peft_sam import PEFT_Sam, AttentionSurgery
+        from micro_sam.v1.models.peft_sam import PEFT_Sam, AttentionSurgery
 
         _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=AttentionSurgery)
         self._check_output(peft_sam)
 
     def test_norm_layer_peft_sam(self):
-        from micro_sam.models.peft_sam import PEFT_Sam, LayerNormSurgery
+        from micro_sam.v1.models.peft_sam import PEFT_Sam, LayerNormSurgery
 
         _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=LayerNormSurgery)
         self._check_output(peft_sam)
 
     def test_bias_layer_peft_sam(self):
-        from micro_sam.models.peft_sam import PEFT_Sam, BiasSurgery
+        from micro_sam.v1.models.peft_sam import PEFT_Sam, BiasSurgery
 
         _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=BiasSurgery)
         self._check_output(peft_sam)
 
     def test_ssf_peft_sam(self):
-        from micro_sam.models.peft_sam import PEFT_Sam, SSFSurgery
+        from micro_sam.v1.models.peft_sam import PEFT_Sam, SSFSurgery
 
         _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=SSFSurgery)
         self._check_output(peft_sam)
 
     def test_adaptformer_peft_sam(self):
-        from micro_sam.models.peft_sam import PEFT_Sam, AdaptFormer
+        from micro_sam.v1.models.peft_sam import PEFT_Sam, AdaptFormer
 
         _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=AdaptFormer, projection_size=64, alpha=2.0, dropout=0.5)
